requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dash==0.36.0
dash-auth==1.3.2
dash-core-components==0.43.0
dash-html-components==0.13.5
dash-renderer==0.17.0
decorator==4.3.2
Flask==1.0.2
Flask-Compress==1.4.0
Flask-SeaSurf==0.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-bigquery==1.8.1
google-cloud-core==0.29.1
google-resumable-media==0.3.2
googleapis-common-protos==1.5.6
gunicorn==19.9.0
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
nbformat==4.4.0
numpy==1.16.0
pandas==0.24.0
plotly==3.5.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
retrying==1.3.3
rsa==4.0
six==1.12.0
traitlets==4.3.2
ua-parser==0.8.0
urllib3==1.24.1
Werkzeug==0.14.1